// src/types/gallery.ts
export interface Gallery {
    id?: number;
    title?: string;
    slug?: string;
    images?: any | null;
    description?: string;
    status?: string;
    createdAt?: Date;
    updatedAt?: Date;
    res?: any;
    result?: any;
}

export interface GalleryFormData {
    id?: number;
    title?: string;
    slug?: string;
    images?: any | null;
    description?: string;
    status?: string;
    createdAt?: Date;
    updatedAt?: Date;
    res?: any;
    result?: any;
}